+++ docs/source/conf.py
@@ -23,7 +23,7 @@ copyright = "2024, Auke Willem Oosterhoff"
 author = "Auke Willem Oosterhoff"
 
 # The full version, including alpha/beta/rc tags
-release = "2.0.0-rc.0"
+release = "2.0.0-rc.2"
 
 
 # -- General configuration ---------------------------------------------------