+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+## 2.0.0-rc.2 (2024-06-18)
+
+- [#315](https://github.com/mobilityhouse/ocpp/pull/315) Allow to skip schema validation in `ChargePoint.call()`. Thanks [@esiebert](https://github.com/esiebert)!
+
+## 2.0.0-rc.1 (2024-06-01)
+
+## BREAKING ##
+- [#642](https://github.com/mobilityhouse/ocpp/pull/642) Fix serializing of "ocpp_csms_url".
+
 ## 2.0.0-rc.0 (2024-05-22)
 
 - [#631](https://github.com/mobilityhouse/ocpp/pull/631) Fix publishing to Pypi.